extracting host name, concating results with filenames

+++ corner_extractor.m
@@ -2,6 +2,7 @@ Files = dir('./frames/out*.png');
 
 BarAppeared = false;
 for File = Files'
+    clearvars CroppedImage CroppedName CroppedTitle;
     try 
         Image = imread(['./frames/' File.name]);
     catch E
@@ -11,8 +12,22 @@ for File = Files'
     CroppedImage = imcrop(Image, [295 452 416 48]);
     GrayCroppedImage = rgb2gray(CroppedImage);
     GrayCroppedImage = imadjust(GrayCroppedImage);
-    CroppedName = imresize(imcrop(GrayCroppedImage, [0 0 416 23]), 2.5);
-    CroppedTitle = imresize(imcrop(GrayCroppedImage, [0 25 416 26]), 2.5);
+    
+    thres = graythresh(GrayCroppedImage);
+    imBW = im2bw(GrayCroppedImage, thres);
+    [L1, L2] = size(GrayCroppedImage);
+    
+    sum = 0;
+    for n = 1:L2
+        sum = sum + imBW(round(L1/2), n);
+    end
+    
+    if(sum == 0)
+        CroppedName = imresize(imcrop(GrayCroppedImage, [0 0 416 23]), 2.5);
+        CroppedTitle = imresize(imcrop(GrayCroppedImage, [0 25 416 26]), 2.5); 
+    else
+        CroppedName = GrayCroppedImage;
+    end
     
     ReducedImage = mean(mean(CroppedImage, 1), 2);
   
@@ -25,7 +40,9 @@ for File = Files'
     if (Dist > 10 && Dist < 50)
         imwrite(CroppedImage, ['./output/cropped-' File.name]);
         imwrite(CroppedName, ['./output/name-' File.name]);
-        imwrite(CroppedTitle, ['./output/title-' File.name]);
+        if (exist('CroppedTitle'))
+            imwrite(CroppedTitle, ['./output/title-' File.name]);
+        end
         disp(['Dist in file ' File.name ' is ' num2str(Dist) '. Image is not empty bar'])
     else
 %         disp('Image is empty bar');
